requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
discord.py==1.6.0
idna==3.1
multidict==5.1.0
numpy==1.19.5
pandas==1.2.1
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
typing-extensions==3.7.4.3
yarl==1.6.3
